requirement.txt
absl-py==2.1.0
ale-py==0.8.1
atari-py==0.2.9
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
glfw==2.8.0
Grid2Op==1.10.5
gymnasium==0.29.1
gymnasium-robotics==1.2.4
huggingface-hub==0.14.1
imageio==2.35.1
imageio-ffmpeg==0.5.1
jupyter==1.1.1
matplotlib==3.6.0
minigrid==3.0.0
miniworld==2.0.0
mpi4py 
mujoco==2.3.3
-e git+https://github.com/openai/mujoco-py@a13903b82f1ab316815e63b3575526005b3b2ae1#egg=mujoco_py
numpy==1.22.4
opencv-python==4.10.0.84
p5py==1.0.0
pandas==2.2.2
patchelf==0.17.2.1
pep517==0.13.1
pillow
protobuf==3.19.6
pytest==8.3.2
requests
rllte-core==1.0.1
sb3_contrib==2.3.0
scikit-learn
scipy==1.13.1
seaborn==0.12.2
stable_baselines3==2.4.1
tensorboard==2.9.1
torch==2.4.0
torchaudio==2.4.0
torchsummary==1.5.1
torchvision==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2

